Don't truncate the top 16 bits of the LE advertisement data mask.
This is stored elsewhere as unsigned 32-bit integer.
It should be stored the same way here.

Bug: 19592970
(cherry picked from commit 91f6987ecca884b0995b77fea014046b8a56385a)

+++ bta/dm/bta_dm_int.h
@@ -515,7 +515,7 @@ typedef struct
 typedef struct
 {
     BT_HDR                  hdr;
-    UINT16                  data_mask;
+    UINT32                  data_mask;
     tBTA_BLE_ADV_DATA       *p_adv_cfg;
     tBTA_SET_ADV_DATA_CMPL_CBACK    *p_adv_data_cback;
 }tBTA_DM_API_SET_ADV_CONFIG;
@@ -1117,4 +1117,3 @@ extern void bta_dm_disable_test_mode(tBTA_DM_MSG *p_data);
 extern void bta_dm_execute_callback(tBTA_DM_MSG *p_data);
 
 #endif /* BTA_DM_INT_H */
-
